requirements.txt
beautifulsoup4==4.9.1
certifi==2018.8.13
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
future==0.16.0
idna==2.7
pycparser==2.20
python-dotenv==0.9.1
python-telegram-bot==12.7
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.24.2
